LayoutTests/fast/dom/MicroData/006-expected.txt

This tests that document.getItems must locate items when parameters are separated by spaces. 
This test also ensure that document.getItems must ignore duplicated tokens.

PASS document.getItems(' http://example.com/foo http://example.com/bar ').length == 2 is true
PASS document.getItems(' http://example.com/foo  data:text/plain  ').length == 2 is true
PASS document.getItems('  http://example.com/foo  data:text/plain  http://example.com/foo').length == 2 is true
PASS successfullyParsed is true

TEST COMPLETE
